Time: 1647 ms (26.48%) | Memory: 249.9 MB (16.59%) - LeetSync

+++ 1087-longest-arithmetic-subsequence/longest-arithmetic-subsequence.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     int longestArithSeqLength(vector<int>& nums) {
-        unordered_map<int,unordered_map<int,long long>> st;
-        long long ans=0;
+        unordered_map<int,unordered_map<int,int >> st;
+        int  ans=0;
         for(int i=0;i<nums.size();i++){
             for(int j=0;j<i;j++){
-                long long diff=(long long)nums[j]-nums[i];
+                int  diff=(int )nums[j]-nums[i];
                 auto it=st[j].find(diff);
                 int count_j=(it==end(st[j]))?1:it->second;
                 st[i][diff]=max(1+count_j,(int)st[i][diff]);
